+++ src/background/state/mod.rs
@@ -7,29 +7,39 @@ use std::collections::HashMap;
 use application::FullState;
 use domain::AhkVar;
 
+use crate::windows_api::monitor::Monitor;
+
 impl FullState {
     pub fn is_weg_enabled(&self) -> bool {
         self.settings.seelenweg.enabled
     }
 
-    pub fn is_weg_enabled_on_monitor(&self, _monitor_idx: usize) -> bool {
-        /* match self.settings.monitors.get(monitor_idx) {
-            Some(monitor) => is_global_enabled && monitor.weg.enabled,
+    pub fn is_weg_enabled_on_monitor(&self, monitor: &Monitor) -> bool {
+        let is_global_enabled = self.is_weg_enabled();
+        let device_id = match monitor.display_device() {
+            Ok(device) => device.id,
+            Err(_) => return is_global_enabled,
+        };
+        match self.settings.monitors_v2.get(&device_id) {
+            Some(config) => is_global_enabled && config.weg.enabled,
             None => is_global_enabled,
-        } */
-        self.is_weg_enabled()
+        }
     }
 
     pub fn is_bar_enabled(&self) -> bool {
         self.settings.fancy_toolbar.enabled
     }
 
-    pub fn is_bar_enabled_on_monitor(&self, _monitor_idx: usize) -> bool {
-        /* match self.settings.monitors.get(monitor_idx) {
-            Some(monitor) => is_global_enabled && monitor.tb.enabled,
+    pub fn is_bar_enabled_on_monitor(&self, monitor: &Monitor) -> bool {
+        let is_global_enabled = self.is_bar_enabled();
+        let device_id = match monitor.display_device() {
+            Ok(device) => device.id,
+            Err(_) => return is_global_enabled,
+        };
+        match self.settings.monitors_v2.get(&device_id) {
+            Some(config) => is_global_enabled && config.tb.enabled,
             None => is_global_enabled,
-        } */
-        self.is_bar_enabled()
+        }
     }
 
     pub fn is_window_manager_enabled(&self) -> bool {
@@ -52,14 +62,18 @@ impl FullState {
         self.settings.ahk_variables.as_hash_map()
     }
 
-    pub fn get_wm_layout_id(&self, _monitor_idx: usize, _workspace_idx: usize) -> String {
-        /* match self.settings.monitors.get(monitor_idx) {
-            Some(monitor) => match monitor.workspaces_v2.get(workspace_idx) {
+    pub fn get_wm_layout_id(&self, monitor: &Monitor, workspace_idx: usize) -> String {
+        let default = self.settings.window_manager.default_layout.clone();
+        let device_id = match monitor.display_device() {
+            Ok(device) => device.id,
+            Err(_) => return default,
+        };
+        match self.settings.monitors_v2.get(&device_id) {
+            Some(config) => match config.workspaces_v2.get(workspace_idx) {
                 Some(workspace) => workspace.layout.clone().unwrap_or(default),
                 None => default,
             },
             None => default,
-        } */
-        self.settings.window_manager.default_layout.clone()
+        }
     }
 }